+++ src/main/java/com/ssginc/login/model/dto/UsersDTO.java
@@ -10,7 +10,7 @@ public class UsersDTO {
     private int usersNo; // 유저번호
     private String usersId; // 유저아이디
     private String usersPw; // 유저비밀번호
-    private String usersRole; // 유저권한
+    private int usersRole; // 유저권한
     private String usersName; // 유저 이름
     private String usersBirth; // 유저생년월일
 }
