+++ src/main/java/com/ssginc/login/service/LoginService.java
@@ -1,12 +1,10 @@
 package com.ssginc.login.service;
 
 import com.ssginc.login.model.dto.UsersDTO;
-import com.ssginc.login.model.dto.UsersDTOTest;
-import com.ssginc.login.model.vo.UsersVO;
 
 public interface LoginService {
     // 회원가입
-    int insertUsers(UsersDTOTest user);
+    int insertUsers(UsersDTO user);
 
     // 아이디 일치 여부 확인
     UsersDTO matchUsersId(String id);
