--- src/main/java/com/ssginc/orders/model/dto/WishOrderDetailsDTO.java
@@ -1,19 +1,0 @@
-package com.ssginc.orders.model.dto;
-
-import com.ssginc.orders.model.vo.ProductsVO;
-import lombok.Builder;
-import lombok.Data;
-
-
-import java.util.List;
-
-@Data
-@Builder
-public class WishOrderDetailsDTO {
-    // 개별 주문 내역 조회 전용 DTO
-    private int orderNo; // 주문 번호
-    private String orderDate; // 주문일자
-    private String userName; // 주문자명
-    private List<WishProductsDTO> products; // 주문 상품 리스트
-    private int totalPrice; // 총 결제 금액
-}
