--- src/main/java/com/ssginc/orders/service/WishOrdersService.java
@@ -1,9 +1,0 @@
-package com.ssginc.orders.service;
-
-import com.ssginc.orders.model.dto.WishProductsDTO;
-
-import java.util.List;
-
-public interface WishOrdersService {
-    int insertOrders(List<WishProductsDTO> products, int usersNo);
-}
